Add baseline game grid & animations
Created base styling for the game board/grid using flexbox. Every row currently color coded for dev purposes (square in the very middle with it's own color). Included is the styling necessary to create our 'X' and 'O' divs as well as their entrance animations for when they appear on the screen. Added is also a special animation for each symbol when the user interacts with them after they appear on the board.